Update date and time consistency
Ensure that date and time are consistently displayed across the dashboard and all widgets. All instances of date and time should match in format and value throughout the application.
[skip gpt_engineer]

+++ src/components/dashboard/AlertTableRow.tsx
@@ -19,7 +19,7 @@ const AlertTableRow = ({ alert, isSelected, onToggle, onTimelineView, visibleCol
     minute: '2-digit',
     second: '2-digit',
     hour12: true,
-    month: 'numeric',
+    month: 'short',
     day: 'numeric',
     year: 'numeric',
   });
